requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
appdirs==1.4.4
    # via black
asgiref==3.3.2
    # via
    #   -c requirements.txt
    #   django
attrs==21.4.0
    # via
    #   -c requirements.txt
    #   black
    #   flake8-bugbear
    #   pytest
    #   syrupy
black==19.10b0
    # via -r requirements-dev.in
certifi==2019.11.28
    # via
    #   -c requirements.txt
    #   requests
    #   urllib3
cffi==1.14.5
    # via
    #   -c requirements.txt
    #   cryptography
chardet==3.0.4
    # via
    #   -c requirements.txt
    #   requests
click==8.0.3
    # via
    #   black
    #   pip-tools
colorama==0.4.4
    # via pytest-watch
colored==1.4.2
    # via syrupy
coreapi==2.3.3
    # via djangorestframework-stubs
coreschema==0.0.4
    # via coreapi
coverage==5.5
    # via pytest-cov
cryptography==37.0.2
    # via
    #   -c requirements.txt
    #   pyopenssl
    #   urllib3
django-stubs-ext==0.2.0
    # via django-stubs
django-stubs==1.8.0
    # via
    #   -r requirements-dev.in
    #   djangorestframework-stubs
django==3.2.12
    # via
    #   -c requirements.txt
    #   django-stubs
    #   django-stubs-ext
djangorestframework-stubs==1.4.0
    # via -r requirements-dev.in
docopt==0.6.2
    # via pytest-watch
fakeredis==1.4.5
    # via -r requirements-dev.in
flake8-bugbear==20.1.4
    # via -r requirements-dev.in
flake8-colors==0.1.6
    # via -r requirements-dev.in
flake8-comprehensions==3.2.2
    # via -r requirements-dev.in
flake8-import-order==0.18.1
    # via -r requirements-dev.in
flake8-logging-format==0.6.0
    # via -r requirements-dev.in
flake8-print==3.1.4
    # via -r requirements-dev.in
flake8==3.9.2
    # via
    #   -r requirements-dev.in
    #   flake8-bugbear
    #   flake8-colors
    #   flake8-comprehensions
    #   flake8-print
flaky==3.7.0
    # via -r requirements-dev.in
freezegun==0.3.15
    # via -r requirements-dev.in
idna==2.8
    # via
    #   -c requirements.txt
    #   requests
    #   urllib3
iniconfig==1.1.1
    # via pytest
isort==5.2.2
    # via -r requirements-dev.in
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
markupsafe==1.1.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via
    #   -c requirements.txt
    #   -r requirements-dev.in
    #   mypy
mypy==0.931
    # via
    #   -r requirements-dev.in
    #   django-stubs
    #   djangorestframework-stubs
numpy==1.21.4
    # via
    #   -c requirements.txt
    #   -r requirements-dev.in
packaging==21.3
    # via
    #   -c requirements.txt
    #   -r requirements-dev.in
    #   pytest
pathspec==0.9.0
    # via black
pip-tools==5.5.0
    # via -r requirements-dev.in
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-import-order
    #   flake8-print
pycparser==2.20
    # via
    #   -c requirements.txt
    #   cffi
pyflakes==2.3.1
    # via flake8
pyopenssl==22.0.0
    # via
    #   -c requirements.txt
    #   urllib3
pyparsing==3.0.7
    # via
    #   -c requirements.txt
    #   packaging
pysocks==1.7.1
    # via
    #   -c requirements.txt
    #   urllib3
pytest-cov==2.12.1
    # via -r requirements-dev.in
pytest-django==4.1.0
    # via -r requirements-dev.in
pytest-env==0.6.2
    # via -r requirements-dev.in
pytest-mock==3.5.1
    # via -r requirements-dev.in
pytest-split==0.6.0
    # via -r requirements-dev.in
pytest-watch==4.2.0
    # via -r requirements-dev.in
pytest==6.2.2
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-mock
    #   pytest-split
    #   pytest-watch
    #   syrupy
python-dateutil==2.8.1
    # via
    #   -c requirements.txt
    #   freezegun
pytz==2021.1
    # via
    #   -c requirements.txt
    #   django
redis==3.4.1
    # via
    #   -c requirements.txt
    #   fakeredis
regex==2022.1.18
    # via black
requests==2.25.1
    # via
    #   -c requirements.txt
    #   coreapi
    #   djangorestframework-stubs
six==1.14.0
    # via
    #   -c requirements.txt
    #   fakeredis
    #   flake8-print
    #   freezegun
    #   python-dateutil
sortedcontainers==2.4.0
    # via
    #   -c requirements.txt
    #   fakeredis
sqlparse==0.3.0
    # via
    #   -c requirements.txt
    #   django
syrupy==1.4.6
    # via -r requirements-dev.in
toml==0.10.1
    # via
    #   black
    #   pytest
    #   pytest-cov
tomli==1.2.3
    # via mypy
typed-ast==1.5.1
    # via black
types-freezegun==1.1.3
    # via -r requirements-dev.in
types-futures==3.3.1
    # via types-protobuf
types-markdown==3.3.9
    # via -r requirements-dev.in
types-protobuf==3.18.2
    # via -r requirements-dev.in
types-python-dateutil==2.8.3
    # via -r requirements-dev.in
types-pytz==2021.3.2
    # via -r requirements-dev.in
types-pyyaml==6.0.1
    # via -r requirements-dev.in
types-redis==4.0.3
    # via -r requirements-dev.in
types-requests==2.26.1
    # via -r requirements-dev.in
typing-extensions==4.1.1
    # via
    #   -c requirements.txt
    #   django-stubs
    #   djangorestframework-stubs
    #   mypy
uritemplate==4.1.1
    # via
    #   -c requirements.txt
    #   coreapi
urllib3[secure,socks]==1.26.5
    # via
    #   -c requirements.txt
    #   requests
watchdog==2.1.8
    # via pytest-watch

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
